feat: 策略模式优惠折扣计算

+++ group-buying-platform-infrastructure/src/main/java/com/vvd/infrastructure/dao/po/GroupBuyDiscount.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GroupBuyDiscount {
-
+    
     /** 自增ID */
     private Long id;
     /** 折扣ID */
@@ -26,7 +26,7 @@ public class GroupBuyDiscount {
     /** 折扣描述 */
     private String discountDesc;
     /** 折扣类型（0:base、1:tag） */
-    private Byte discountType;
+    private Integer discountType;
     /** 营销优惠计划（ZJ:直减、MJ:满减、N元购） */
     private String marketPlan;
     /** 营销优惠表达式 */
@@ -37,6 +37,6 @@ public class GroupBuyDiscount {
     private Date createTime;
     /** 更新时间 */
     private Date updateTime;
-
+    
 }
